Resolve the Callbacks cycle, option 1
