+++ src/callbacks_test.ts
@@ -15,18 +15,12 @@
 // tslint:disable:max-line-length
 import {scalar} from '@tensorflow/tfjs-core';
 
-import {BaseLogger, CallbackList, History} from './callbacks';
+import {BaseLogger, CallbackList} from './callbacks';
 import {disposeTensorsInLogs, resolveScalarsInLogs, UnresolvedLogs} from './engine/logs';
-import {Model} from './engine/training';
+import {History} from './engine/training';
 import {describeMathCPUAndGPU} from './utils/test_utils';
 
 // tslint:enable:max-line-length
-
-class MockModel extends Model {
-  constructor(name: string) {
-    super({inputs: [], outputs: [], name});
-  }
-}
 
 describe('BaseLogger Callback', () => {
   it('Records and averages losses in an epoch', async done => {
@@ -73,28 +67,6 @@ describe('BaseLogger Callback', () => {
   });
 });
 
-describe('History Callback', () => {
-  it('onTrainBegin', async done => {
-    const history = new History();
-    await history.onTrainBegin();
-    expect(history.epoch).toEqual([]);
-    expect(history.history).toEqual({});
-    done();
-  });
-  it('onEpochEnd', async done => {
-    const history = new History();
-    await history.onTrainBegin();
-    await history.onEpochEnd(0, {'val_loss': 10, 'val_accuracy': 0.1});
-    expect(history.epoch).toEqual([0]);
-    expect(history.history).toEqual({'val_loss': [10], 'val_accuracy': [0.1]});
-    await history.onEpochEnd(1, {'val_loss': 9.5, 'val_accuracy': 0.2});
-    expect(history.epoch).toEqual([0, 1]);
-    expect(history.history)
-        .toEqual({'val_loss': [10, 9.5], 'val_accuracy': [0.1, 0.2]});
-    done();
-  });
-});
-
 describe('CallbackList', () => {
   it('Constructor with empty arg', async done => {
     const callbackList = new CallbackList();
@@ -110,15 +82,6 @@ describe('CallbackList', () => {
     callbackList.setParams(params);
     expect(history1.params).toEqual(params);
     expect(history2.params).toEqual(params);
-  });
-  it('Constructor and setModel with array of callbacks', () => {
-    const history1 = new History();
-    const history2 = new History();
-    const callbackList = new CallbackList([history1, history2]);
-    const model = new MockModel('MockModelA');
-    callbackList.setModel(model);
-    expect(history1.model).toEqual(model);
-    expect(history2.model).toEqual(model);
   });
   it('onTrainBegin', async done => {
     const history1 = new History();
